refactor: consolidate redundant commands (phase 2)
Remove duplicate commands to simplify the CLI:

- Remove `worker` command (use `work --mode=redis` instead)
- Remove `agent` command (was deprecated, use `work --mode=nexus`)
- Remove `restart` command (use `run --restart` instead)
- Remove `join` command (use `login --authkey` instead)

Changes:
- Add `--restart` flag to `run` command
- Add `--authkey` and `--node-name` flags to `login` command
- Create `job_handlers.go` for job execution code used by test.go
- Update init.go to use `login --authkey` instead of `join`
- Update documentation and hint messages

Final command list:
  init, login, logout, run, stop, status, logs, test, nodes, work,
  expose, version, completion
